build(deps)!: pin libGravity at 3254fc3, the integer tempo arithmetic
The fork replaces the float tempo arithmetic of the vendored uClock with an
integer one. Measured on this firmware: Flash 28764 to 27672, so -1092 bytes,
and RAM +2. The float runtime of avr-libc leaves the image entirely, 1474 bytes
over 23 symbols.

ADR 0008 authorises it since its amendment of 2026-09-03, under thirteen
conditions, and the twelve that can be verified before the bump were verified in
lot S2.1. The public interface of libGravity does not change: the only callers
of the tempo API are three integer lines of src/clock.h.

The characterization suite carries one line of the bump. test_gravity.cpp
DEFINES a stub uClockClass::setTempo, so the signature had to follow. It is the
only stub of the whole test tree that touches the tempo API, and the suite stays
conform: seven audited anomalies, no more and no fewer.

+++ test/test_gravity/test_gravity.cpp
@@ -41,7 +41,7 @@ uClockClass::ClockMode uClockClass::getClockMode() {
     return clock_mode;
 }
 
-void uClockClass::setTempo(float tempo) {
+void uClockClass::setTempo(uint16_t tempo) {
     (void)tempo;
 }
